add function component
