+++ src/components/TodoList/TodoList.jsx
@@ -1,20 +1,13 @@
 import React, { useState } from 'react';
+import AddTodo from '../AddTodo/AddTodo';
 
 export default function TodoList() {
   const [todos, setTodos] = useState([
     { id: '123', text: '장보기', status: 'active' },
     { id: '124', text: '공부하기', status: 'active' },
   ]);
-  const [text, setText] = useState('');
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    setTodos([...todos, { id: '125', text, status: 'active' }]);
-    setText('');
-  };
 
-  const handleChange = (e) => {
-    setText(e.target.value);
-  };
+  const handleAdd = (todo) => setTodos([...todos, todo]);
 
   return (
     <section>
@@ -25,15 +18,7 @@ export default function TodoList() {
           </li>
         ))}
       </ul>
-      <form onSubmit={handleSubmit}>
-        <input
-          type='text'
-          placeholder='type your todo'
-          onChange={handleChange}
-          value={text}
-        />
-        <button>Add</button>
-      </form>
+      <AddTodo onAdd={handleAdd} />
     </section>
   );
 }
